Move languages out of MainContext (#20503)
* move languages off of MainContext

* load languages from request to workaround sharing the languages file
