+++ components/Search.tsx
@@ -6,6 +6,7 @@ import { sendEvent, EventType } from 'components/lib/events'
 import { useMainContext } from './context/MainContext'
 import { useVersion } from 'components/hooks/useVersion'
 import cx from 'classnames'
+import { useLanguages } from './context/LanguagesContext'
 
 type SearchResult = {
   url: string
@@ -30,9 +31,10 @@ export function Search({ isStandalone = false, updateSearchParams = true, childr
   const inputRef = useRef<HTMLInputElement>(null)
   const { t } = useTranslation('search')
   const { currentVersion } = useVersion()
+  const { languages } = useLanguages()
 
   // Figure out language and version for index
-  const { languages, searchVersions, nonEnterpriseDefaultVersion } = useMainContext()
+  const { searchVersions, nonEnterpriseDefaultVersion } = useMainContext()
   // fall back to the non-enterprise default version (FPT currently) on the homepage, 404 page, etc.
   const version = searchVersions[currentVersion] || searchVersions[nonEnterpriseDefaultVersion]
   const language = (Object.keys(languages).includes(router.locale || '') && router.locale) || 'en'
