+++ components/context/MainContext.tsx
@@ -13,14 +13,6 @@ type ProductT = {
   versions?: Array<string>
 }
 
-type LanguageItem = {
-  name: string
-  nativeName: string
-  code: string
-  hreflang: string
-  wip?: boolean
-}
-
 type VersionItem = {
   version: string
   versionTitle: string
@@ -83,7 +75,6 @@ export type MainContextT = {
   currentPathWithoutLanguage: string
   currentLanguage: string
   userLanguage: string
-  languages: Record<string, LanguageItem>
   allVersions: Record<string, VersionItem>
   currentProductTree?: ProductTreeNode | null
   featureFlags: FeatureFlags
@@ -162,20 +153,6 @@ export const getMainContextFromRequest = (req: any): MainContextT => {
     enterpriseServerVersions: req.context.enterpriseServerVersions,
     currentLanguage: req.context.currentLanguage,
     userLanguage: req.context.userLanguage || '',
-    languages: Object.fromEntries(
-      Object.entries(req.context.languages).map(([key, entry]: any) => {
-        return [
-          key,
-          {
-            name: entry.name,
-            nativeName: entry.nativeName || '',
-            code: entry.code,
-            hreflang: entry.hreflang,
-            wip: entry.wip || false,
-          },
-        ]
-      })
-    ),
     allVersions: req.context.allVersions,
     currentProductTree: req.context.currentProductTree
       ? getCurrentProductTree(req.context.currentProductTree)
